Add login / signup rpcs
